Updated Sil Model, Bill parser, and Sil.java
Removed Model classes copied from legislativo-api-java project. Instead of
that, a jar with the model is included in the project.

The parser uses temporal objects (SilBill for Bill) that uses
legislativo-api.model clases and additional attributes that are needed to
finally build the complete entity.

The JAR should be updated everytime the model is modified.